new incoming type `float`, old was `int` , in result was rough jumping in sizes after updateViewport

+++ gltf/src/net/mgsx/gltf/scene3d/scene/SceneManager.java
@@ -265,7 +265,7 @@ public class SceneManager implements Disposable {
 		return renderableProviders;
 	}
 
-	public void updateViewport(int width, int height) {
+	public void updateViewport(float width, float height) {
 		if(camera != null){
 			camera.viewportWidth = width;
 			camera.viewportHeight = height;
